refactor(agent): own experiment runtime
Move the repository-work runner and container adapter out of generic source tooling into the independently owned experiment framework while preserving the Shifu command surface.
