+++ package.json
@@ -127,7 +127,7 @@
     "test:run-agent-opencode-continuity": "node scripts/require-shifu.mjs test:run-agent-opencode-continuity && node --test scripts/run-agent-opencode-continuity.test.mjs",
     "qualify:run-agent-opencode-continuity": "node scripts/require-shifu.mjs qualify:run-agent-opencode-continuity && node scripts/run-agent-opencode-continuity.mjs",
     "test:agent-repository-work": "node scripts/require-shifu.mjs test:agent-repository-work && node --test scripts/run-agent-repository-work-experiment.test.mjs scripts/agent-repository-work-workflow.test.mjs",
-    "qualify:agent-repository-work": "node scripts/require-shifu.mjs qualify:agent-repository-work && node scripts/run-agent-repository-work-experiment.mjs",
+    "qualify:agent-repository-work": "node scripts/require-shifu.mjs qualify:agent-repository-work && node framework/agent-repository-work/run.mjs",
     "check:agent-pack": "node scripts/require-shifu.mjs check:agent-pack && node scripts/verify-agent-pack.mjs",
     "check:fact-cut-kernel-contract": "node scripts/require-shifu.mjs check:fact-cut-kernel-contract && node --test scripts/check-fact-cut-kernel-contract.test.mjs",
     "check:work-lifecycle-operation-matrix": "node scripts/require-shifu.mjs check:work-lifecycle-operation-matrix && node scripts/materialize-work-lifecycle-operation-matrix.mjs --check && node scripts/render-work-lifecycle-operation-matrix.mjs --check && node developer/sdk/src/sdk.js contract render work-lifecycle-operation-matrix --check --json && node developer/sdk/src/sdk.js contract policy --check --json && node developer/sdk/src/sdk.js contract audit --json && node --test scripts/check-work-lifecycle-operation-matrix.test.mjs scripts/check-native-authority-membrane.test.mjs",
